added volume compute method

+++ domain/src/main/java/ru/fastdelivery/domain/common/dimesions/Length.java
@@ -1,10 +1,8 @@
 package ru.fastdelivery.domain.common.dimesions;
-
-import java.math.BigInteger;
 /**
- * Размерность для опредиления габаритов в миллиметрах
+ * Размерность для определения габаритов в миллиметрах
  */
 public record Length(
-        BigInteger length
+        Integer length
 ) {
 }
